+++ packages/ricos-modals/web/src/components/Toolbar.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import type { ReactNode } from 'react';
 import type { ModalConfig } from 'ricos-types';
-import styles from '../../statics/styles/popover.scss';
+import styles from '../../statics/styles/toolbar.scss';
 import { Popper } from './Popper';
 
 interface Props {
@@ -10,9 +10,9 @@ interface Props {
   closeModal: () => void;
 }
 
-export const Popover = ({ children, modalConfig, closeModal }: Props) => {
+export const Toolbar = ({ children, modalConfig, closeModal }: Props) => {
   return (
-    <Popper className={styles.popover} modalConfig={modalConfig} closeModal={closeModal}>
+    <Popper className={styles.toolbar} modalConfig={modalConfig} closeModal={closeModal}>
       {children}
     </Popper>
   );